Revert "doom doomdoom doom doom!"
This reverts commit cdede66c2f6a03714d965dad53331d004315fc6b.
It was a bad commit. Girl troubles. Hate to see it.
Been working on this but I'm pretty sure flattening the tables to make this work nulls the point of threading, so:
Probably just going to see if I can get away with only threading the write(ostring) call and move on.
Will work on optimizations when I have worse hardware to test my table-write on.